test/requirements.txt
# Python requirements for unit tests
-r ../requirements.txt

# alignak - develop+environment branch
-e git+https://github.com/Alignak-monitoring/alignak.git@python3#egg=alignak

# alignak backend
-e git+https://github.com/Alignak-monitoring-contrib/alignak-backend.git@develop#egg=alignak-backend

# alignak backend import
-e git+https://github.com/Alignak-monitoring-contrib/alignak-backend-import.git@develop#egg=alignak-backend-import

unittest2
requests_mock

# Tests time freeze
freezegun

# System monitor
psutil

# Use py.test as test-runner
pytest
pytest-cov
# Makes some coverage report for the tests
coverage
# Report coverage results to coveralls.io
coveralls

# Static code and doc analysis libraries
pylint
pycodestyle
pep257

# UWSGI Pyphton Web application server
uwsgi
